+++ paralel/Personal_Tiles/src/puzzle/Freelances.java
@@ -1,0 +1,30 @@
+//package puzzle;
+//
+//public class Freelances extends Tiles{
+//    private int posX;
+//    private int posY;
+//    private String name;
+//    private String color;
+//    private Rectangle tile;
+//    private boolean visible;
+//
+//    public Freelances(int posX, int posY, String color) {
+//        super(posX, posY, color);
+//        makeVisibleCreate();
+//    }
+//    public void makeVisibleCreate() {
+//        tile.changeColor(this.color);
+//        tile.changeSize(20, 20);
+//        tile.moveHorizontal(posX * 23);
+//        tile.moveVertical(posY * 23);
+//        tile.makeVisible();
+//    }
+//    public void makeVisible() {
+//        tile.makeVisible();
+//        this.visible = true;
+//    }
+//    public void makeInvisible() {
+//        tile.makeInvisible();
+//        this.visible = false;
+//    }
+//}
