+++ paralel/Personal_Tiles/src/puzzle/Glues.java
@@ -1,0 +1,4 @@
+//package puzzle;
+//
+//public class Glues {
+//}
