+++ paralel/Personal_Tiles/src/puzzle/PuzzleContest.java
@@ -1,3 +1,5 @@
+package puzzle;
+
 import java.util.ArrayList;
 import java.util.Arrays;
