+++ paralel/Personal_Tiles/src/shapes/Circle.java
@@ -1,81 +1,94 @@
+package shapes;
+
 import java.awt.*;
+import java.awt.geom.*;
 
 /**
- * A triangle that can be manipulated and that draws itself on a canvas.
+ * A circle that can be manipulated and that draws itself on a canvas.
  *
  * @author  Michael Kolling and David J. Barnes
- * @version 1.0  (15 July 2000)
+ * @version 1.0.  (15 July 2000)
  */
 
-public class Triangle{
+public class Circle{
 
-    public static int VERTICES=3;
+    public static final double PI=3.1416;
 
-    private int height;
-    private int width;
+    private int diameter;
     private int xPosition;
     private int yPosition;
     private String color;
     private boolean isVisible;
 
-    /**
-     * Create a new triangle at default position with default color.
-     */
-    public Triangle(){
-        height = 30;
-        width = 40;
-        xPosition =80;
+
+    public Circle(){
+        diameter = 30;
+        xPosition = 20;
         yPosition = 15;
-        color = "green";
+        color = "blue";
         isVisible = false;
     }
 
-    /**
-     * Make this triangle visible. If it was already visible, do nothing.
-     */
+
+
     public void makeVisible(){
         isVisible = true;
         draw();
     }
 
-    /**
-     * Make this triangle invisible. If it was already invisible, do nothing.
-     */
+
     public void makeInvisible(){
         erase();
         isVisible = false;
     }
 
+    private void draw(){
+        if(isVisible) {
+            Canvas canvas = Canvas.getCanvas();
+            canvas.draw(this, color,
+                    new Ellipse2D.Double(xPosition, yPosition,
+                            diameter, diameter));
+            canvas.wait(10);
+        }
+    }
+
+    private void erase(){
+        if(isVisible) {
+            Canvas canvas = Canvas.getCanvas();
+            canvas.erase(this);
+        }
+    }
+
     /**
-     * Move the triangle a few pixels to the right.
+     * Move the circle a few pixels to the right.
      */
     public void moveRight(){
         moveHorizontal(20);
     }
 
     /**
-     * Move the triangle a few pixels to the left.
+     * Move the circle a few pixels to the left.
      */
     public void moveLeft(){
         moveHorizontal(-20);
     }
 
     /**
-     * Move the triangle a few pixels up.
+     * Move the circle a few pixels up.
      */
     public void moveUp(){
         moveVertical(-20);
     }
 
     /**
-     * Move the triangle a few pixels down.
+     * Move the circle a few pixels down.
      */
     public void moveDown(){
         moveVertical(20);
     }
 
     /**
-     * Move the triangle horizontally.
+     * Move the circle horizontally.
      * @param distance the desired distance in pixels
      */
     public void moveHorizontal(int distance){
@@ -85,7 +98,7 @@ public class Triangle{
     }
 
     /**
-     * Move the triangle vertically.
+     * Move the circle vertically.
      * @param distance the desired distance in pixels
      */
     public void moveVertical(int distance){
@@ -95,7 +108,7 @@ public class Triangle{
     }
 
     /**
-     * Slowly move the triangle horizontally.
+     * Slowly move the circle horizontally.
      * @param distance the desired distance in pixels
      */
     public void slowMoveHorizontal(int distance){
@@ -115,7 +128,7 @@ public class Triangle{
     }
 
     /**
-     * Slowly move the triangle vertically.
+     * Slowly move the circle vertically
      * @param distance the desired distance in pixels
      */
     public void slowMoveVertical(int distance){
@@ -124,7 +137,7 @@ public class Triangle{
         if(distance < 0) {
             delta = -1;
             distance = -distance;
-        } else {
+        }else {
             delta = 1;
         }
 
@@ -135,14 +148,12 @@ public class Triangle{
     }
 
     /**
-     * Change the size to the new size
-     * @param newHeight the new height in pixels. newHeight must be >=0.
-     * @param newWidht the new width in pixels. newWidht must be >=0.
+     * Change the size.
+     * @param newDiameter the new size (in pixels). Size must be >=0.
      */
-    public void changeSize(int newHeight, int newWidth) {
+    public void changeSize(int newDiameter){
         erase();
-        height = newHeight;
-        width = newWidth;
+        diameter = newDiameter;
         draw();
     }
 
@@ -156,26 +167,6 @@ public class Triangle{
         draw();
     }
 
-    /*
-     * Draw the triangle with current specifications on screen.
-     */
-    private void draw(){
-        if(isVisible) {
-            Canvas canvas = Canvas.getCanvas();
-            int[] xpoints = { xPosition, xPosition + (width/2), xPosition - (width/2) };
-            int[] ypoints = { yPosition, yPosition + height, yPosition + height };
-            canvas.draw(this, color, new Polygon(xpoints, ypoints, 3));
-            canvas.wait(10);
-        }
-    }
 
-    /*
-     * Erase the triangle on screen.
-     */
-    private void erase(){
-        if(isVisible) {
-            Canvas canvas = Canvas.getCanvas();
-            canvas.erase(this);
-        }
-    }
+
 }
