+++ paralel/Personal_Tiles/src/shapes/Rectangle.java
@@ -1,3 +1,5 @@
+package shapes;
+
 /**
  * A rectangle that can be manipulated and that draws itself on a canvas.
  *
